separating tests and modularising tag manager script
