+++ app/views/includes/ga_head.njk
@@ -1,0 +1,28 @@
+<!-- Google Tag Manager -->
+{% if cmUserPreferences['analytics'] === 'on' %}
+  <script>
+    dataLayer = [];
+  </script>
+  <script>
+    (function (w, d, s, l, i) {
+      w[l] = w[l] || [];
+      w[l].push({'gtm.start': new Date().getTime(), event: 'gtm.js'});
+      var f = d.getElementsByTagName(s)[0],
+        j = d.createElement(s),
+        dl = l != 'dataLayer'
+          ? '&l=' + l
+          : '';
+      j.async = true;
+      j.src = 'https://www.googletagmanager.com/gtm.js?id=' + i + dl;
+      f
+        .parentNode
+        .insertBefore(j, f);
+    })(window, document, 'script', 'dataLayer', '{{ GOOGLE_ANALYTICS_ID }}');
+  </script>
+  <script>
+    function gtag() {
+      dataLayer.push(arguments)
+    }
+  </script>
+{% endif %}
+<!-- End Google Tag Manager -->
