+++ app/views/includes/head.njk
@@ -1,3 +1,4 @@
+{% include "./ga_head.njk" %}
 <link href="{{ css_path }}" media="screen" rel="stylesheet" type="text/css"/>
 <link href="{{ css_path }}" media="print" rel="stylesheet" type="text/css"/>
 <style>
@@ -8,31 +9,3 @@
     display: none !important;
   }
 </style>
-<script>
-  dataLayer = [];
-</script>
-<!-- Google Tag Manager -->
-{% if cmUserPreferences['analytics'] === 'on' %}
-  <script>
-    (function (w, d, s, l, i) {
-      w[l] = w[l] || [];
-      w[l].push({'gtm.start': new Date().getTime(), event: 'gtm.js'});
-      var f = d.getElementsByTagName(s)[0],
-        j = d.createElement(s),
-        dl = l != 'dataLayer'
-          ? '&l=' + l
-          : '';
-      j.async = true;
-      j.src = 'https://www.googletagmanager.com/gtm.js?id=' + i + dl;
-      f
-        .parentNode
-        .insertBefore(j, f);
-    })(window, document, 'script', 'dataLayer', '{{ GOOGLE_ANALYTICS_ID }}');
-  </script>
-{% endif %}
-<!-- End Google Tag Manager -->
-<script>
-  function gtag() {
-    dataLayer.push(arguments)
-  }
-</script>
